# api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Mall API
  version: 1.1.0
  description: 商城商品与认证 API；本文件是前后端并行开发的唯一接口契约
x-development-conventions:
  code-comments:
    required: true
    scope: frontend-and-backend
    rule: 关键业务逻辑、认证流程、安全校验、异常分支和复杂算法必须添加中文注释，说明设计意图；避免只复述代码表面行为。
servers:
  - url: http://localhost:8080
tags:
  - name: Auth
  - name: Shop
  - name: Products
paths:
  /api/v1/shop/products:
    get:
      tags: [Shop]
      summary: 消费者分页浏览已上架商品
      operationId: listShopProducts
      security: []
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: keyword, in: query, description: 按商品名称搜索, schema: { type: string, maxLength: 200 } }
      responses:
        '200':
          description: 查询成功
          content: { application/json: { schema: { $ref: '#/components/schemas/ShopProductPage' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/shop/products/{id}:
    get:
      tags: [Shop]
      summary: 获取已上架商品的消费者详情
      operationId: getShopProduct
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, pattern: '^[1-9][0-9]*$' }
      responses:
        '200':
          description: 查询成功
          content: { application/json: { schema: { $ref: '#/components/schemas/ShopProductDetail' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: 注册商城账号
      operationId: register
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RegisterRequest' } } }
      responses:
        '201':
          description: 注册成功
          content: { application/json: { schema: { $ref: '#/components/schemas/CurrentUser' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: 用户名已存在
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 账号密码登录
      operationId: login
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } }
      responses:
        '200':
          description: 登录成功
          content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '423': { $ref: '#/components/responses/AccountLocked' }
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: 刷新访问令牌（刷新令牌单次使用）
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RefreshTokenRequest' } } }
      responses:
        '200':
          description: 刷新成功，同时轮换 accessToken 和 refreshToken
          content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: 退出当前会话
      operationId: logout
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RefreshTokenRequest' } } }
      responses:
        '204': { description: 已退出；重复调用也返回 204 }
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: 获取当前登录用户
      operationId: getCurrentUser
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 查询成功
          content: { application/json: { schema: { $ref: '#/components/schemas/CurrentUser' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/products:
    get:
      tags: [Products]
      summary: 分页查询商品
      operationId: listProducts
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: keyword, in: query, description: 按名称或 SKU 搜索, schema: { type: string, maxLength: 200 } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ProductStatus' } }
      responses:
        '200':
          description: 查询成功
          content: { application/json: { schema: { $ref: '#/components/schemas/ProductPage' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Products]
      summary: 新增商品
      operationId: createProduct
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProductWriteRequest' } } }
      responses:
        '201':
          description: 创建成功
          content: { application/json: { schema: { $ref: '#/components/schemas/ProductDetail' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, pattern: '^[1-9][0-9]*$' }
    get:
      tags: [Products]
      summary: 获取商品详情
      operationId: getProduct
      responses:
        '200':
          description: 查询成功
          content: { application/json: { schema: { $ref: '#/components/schemas/ProductDetail' } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Products]
      summary: 编辑商品
      operationId: updateProduct
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProductWriteRequest' } } }
      responses:
        '200':
          description: 更新成功
          content: { application/json: { schema: { $ref: '#/components/schemas/ProductDetail' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Products]
      summary: 删除商品
      operationId: deleteProduct
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: 删除成功 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ShopProductListItem:
      type: object
      additionalProperties: false
      required: [id, name, price, soldOut]
      properties:
        id: { type: string, example: '1' }
        name: { type: string, example: 示例商品 }
        price: { type: number, multipleOf: 0.01, example: 99.00 }
        imageUrl: { type: string, format: uri, nullable: true }
        soldOut: { type: boolean, description: 是否售罄 }
    ShopProductDetail:
      allOf:
        - { $ref: '#/components/schemas/ShopProductListItem' }
        - type: object
          properties:
            description: { type: string, nullable: true }
    ShopProductPage:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total, totalPages]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ShopProductListItem' } }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
        totalPages: { type: integer, format: int64 }
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [username, password, displayName]
      properties:
        username: { type: string, minLength: 3, maxLength: 64, pattern: '^[A-Za-z0-9_]+$', example: mall_user }
        password: { type: string, minLength: 8, maxLength: 72, format: password, example: User123! }
        displayName: { type: string, minLength: 1, maxLength: 100, example: 商城用户 }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 64, example: admin }
        password: { type: string, minLength: 8, maxLength: 72, format: password, example: Admin123! }
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken: { type: string, minLength: 32 }
    TokenResponse:
      type: object
      additionalProperties: false
      required: [tokenType, accessToken, expiresIn, refreshToken, refreshExpiresIn, user]
      properties:
        tokenType: { type: string, enum: [Bearer] }
        accessToken: { type: string, description: 短期 JWT }
        expiresIn: { type: integer, format: int64, example: 900, description: accessToken 剩余有效秒数 }
        refreshToken: { type: string, description: 不透明的单次使用刷新令牌 }
        refreshExpiresIn: { type: integer, format: int64, example: 604800, description: refreshToken 剩余有效秒数 }
        user: { $ref: '#/components/schemas/CurrentUser' }
    CurrentUser:
      type: object
      additionalProperties: false
      required: [id, username, displayName, roles]
      properties:
        id: { type: string, example: '1' }
        username: { type: string, example: admin }
        displayName: { type: string, example: 商城管理员 }
        avatarUrl: { type: string, format: uri, nullable: true }
        roles:
          type: array
          uniqueItems: true
          items: { type: string, enum: [ADMIN, OPERATOR] }
    ProductStatus:
      type: integer
      enum: [0, 1]
      description: 0 下架，1 上架
    ProductWriteRequest:
      type: object
      required: [sku, name, price, stock, status]
      properties:
        sku: { type: string, minLength: 1, maxLength: 64, example: PHONE-001 }
        name: { type: string, minLength: 1, maxLength: 200, example: 示例商品 }
        price: { type: number, minimum: 0, multipleOf: 0.01, example: 1999.00 }
        stock: { type: integer, minimum: 0, maximum: 2147483647, example: 100 }
        status: { $ref: '#/components/schemas/ProductStatus' }
        imageUrl: { type: string, maxLength: 1000, nullable: true }
        description: { type: string, nullable: true }
    ProductListItem:
      type: object
      required: [id, sku, name, price, stock, status, createdAt, updatedAt]
      properties:
        id: { type: string, example: '1' }
        sku: { type: string }
        name: { type: string }
        price: { type: number }
        stock: { type: integer }
        status: { $ref: '#/components/schemas/ProductStatus' }
        imageUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProductDetail:
      allOf:
        - { $ref: '#/components/schemas/ProductListItem' }
        - type: object
          properties: { description: { type: string, nullable: true } }
    ProductPage:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ProductListItem' } }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
        totalPages: { type: integer, format: int64 }
    FieldError:
      type: object
      properties: { field: { type: string }, message: { type: string } }
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code: { type: string, example: PRODUCT_NOT_FOUND }
        message: { type: string, example: 商品不存在 }
        details: { type: array, nullable: true, items: { $ref: '#/components/schemas/FieldError' } }
        timestamp: { type: string, format: date-time }
  responses:
    BadRequest:
      description: 请求参数错误
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    NotFound:
      description: 商品不存在
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Conflict:
      description: SKU 已存在
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    InternalError:
      description: 服务端异常
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Unauthorized:
      description: 凭证错误、缺失、过期或已撤销
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Forbidden:
      description: 已登录但权限不足或账号被禁用
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    AccountLocked:
      description: 登录失败次数过多，账号暂时锁定
      headers:
        Retry-After: { schema: { type: integer }, description: 建议等待秒数 }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }